requirements.txt
fastmcp>=2.12.0
requests>=2.32.0
python-dotenv>=1.0.0
boto3>=1.34.0
# Optional: For enhanced authentication features
# cryptography>=41.0.0  # For JWT token support if needed
# pyjwt>=2.8.0          # For JWT token validation if needed
